started implementing FirmwareDevice
fixed some typos

+++ X86_64_EMU_SOFT/HEADER/X86_64_EMU_SOFT/HELPERS/SET_BIT.h
@@ -14,10 +14,10 @@ namespace X86_64_EMU_SOFT
 			return (value & ~(1ULL << bit)) | (uint32_t((set ? 1ULL : 0ULL)) << bit);
 		}
 		[[nodiscard]] inline uint16_t SetBit(uint16_t value, uint8_t bit, bool set) noexcept {
-			return (value & ~(uint16_t(1) << bit)) | (uint16_t((set ? 1ULL : 0ULL)) << bit);
+			return static_cast< uint16_t>( (value & ~(uint16_t(1) << bit)) | (uint16_t((set ? 1ULL : 0ULL)) << bit));
 		}
 		[[nodiscard]] inline uint8_t SetBit(uint8_t value, uint8_t bit, bool set) noexcept {
-			return (value & ~(1ull << bit)) | (uint8_t((set ? 1ULL : 0ULL)) << bit);
+			return static_cast<uint8_t>((value & ~(1ull << bit)) | (uint8_t((set ? 1ULL : 0ULL)) << bit));
 		}
 	}
 }
